move Revise dep to an weakdep
This makes Revise a proper pkgextension/weakdep. This should
have no impact to functionality, compiletime, or runtime.
Though it should reduce direct dependency count at precompile time.
